+++ src/main/java/com/github/susonicth/advent/advent04/Advent0401.java
@@ -5,7 +5,7 @@ import com.github.susonicth.advent.Advent;
 public class Advent0401 extends Advent {
 
     @Override
-    protected int mapLineToNumber(String line) {
+    protected long mapLineToNumber(String line) {
         return new Card(line).getScore();
     }
